Rename qiskit._util
