+++ qiskit/providers/basicaer/statevector_simulator.py
@@ -22,7 +22,7 @@ The input qobj to this simulator has no shots, no measures, no reset, no noise.
 
 import logging
 from math import log2
-from qiskit._util import local_hardware_info
+from qiskit.util import local_hardware_info
 from qiskit.providers.basicaer.exceptions import BasicAerError
 from qiskit.providers.models import BackendConfiguration
 from .qasm_simulator import QasmSimulatorPy
