+++ qiskit/tools/parallel.py
@@ -47,7 +47,7 @@ import os
 import platform
 from multiprocessing import Pool
 from qiskit.exceptions import QiskitError
-from qiskit._util import local_hardware_info
+from qiskit.util import local_hardware_info
 from qiskit.tools.events._pubsub import Publisher
 
 # Set parallel flag
