+++ qiskit/util.py
